refactor: TestDataSetUpHelper를 BaseIntegrationTest로 통합
